graphics clipping fix(hopefully)

+++ IPDefinitiveVersion/API_Headers/IGraphics_API.h
@@ -9,12 +9,12 @@ public:
 	virtual void setCameraEntity(uint32_t) = 0;
 	virtual void setFOV(uint32_t) = 0;
 	virtual void setResolution(uint32_t, uint32_t) = 0;
-	virtual void setClipRange(uint32_t, uint32_t) = 0;
+	virtual void setClipRange(float, float) = 0;
 
 	virtual uint32_t getCameraEntity() = 0;
 	virtual uint32_t getFOV() = 0;
 	virtual void getResolution(uint32_t&, uint32_t&) = 0;
-	virtual void getClipRange(uint32_t&, uint32_t&) = 0;
+	virtual void getClipRange(float&, float&) = 0;
 
 	struct renderMatrixes
 	{
